# openapi/flora.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "flora API",
    "description": "Manages @Flora guild deployments, User Bots, Server Custom Bots, authentication, builds, secrets, key-value stores, logs, and operator orchestration.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "flora API base path"
    }
  ],
  "paths": {
...
    "/custom-bots/{bot_id}/deployment": {
      "get": {
        "tags": ["User Bots"],
        "summary": "Get a User Bot deployment",
        "description": "Returns the active deployment source and bundle for a User Bot owned by the authenticated user. Returns a not-found problem when the bot has not been deployed.",
        "operationId": "get_custom_bot_deployment",
        "parameters": [
          {
            "name": "bot_id",
            "in": "path",
            "description": "User Bot ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Active User Bot deployment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomBotDeploymentResponse"
                }
              }
            }
          },
          "default": {
            "description": "The request failed. Inspect the HTTP status and Problem Details response for recovery guidance.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["User Bots"],
        "summary": "Deploy a User Bot",
        "description": "Builds or accepts the supplied bundle, atomically replaces the User Bot isolate, records a deployment revision, and starts its Discord gateway. If deployment fails, the previous runtime and persisted deployment are restored.",
        "operationId": "deploy_custom_bot",
        "parameters": [
          {
            "name": "bot_id",
            "in": "path",
            "description": "User Bot ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeploymentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Active User Bot deployment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomBotDeploymentResponse"
                }
              }
            }
          },
          "default": {
            "description": "The request failed. Inspect the HTTP status and Problem Details response for recovery guidance.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/deployments/": {
      "get": {
        "tags": ["Deployments"],
        "summary": "List deployments",
        "description": "Returns all deployment snapshots the authenticated user has access to.",
        "operationId": "list_deployments_handler",
        "responses": {
          "200": {
            "description": "Deployments list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeploymentListItem"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "default": {
            "description": "The request failed. Inspect the HTTP status and Problem Details response for recovery guidance.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/deployments/{guild_id}": {
      "get": {
        "tags": ["Deployments"],
        "summary": "Get deployment",
        "description": "Returns the latest successful deployment snapshot for a guild.",
        "operationId": "get_deployment_handler",
        "parameters": [
          {
            "name": "guild_id",
            "in": "path",
            "description": "Guild ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "include_bundle",
            "in": "query",
            "description": "Include bundled output in response",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deployment found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentResponse"
                }
              }
            }
          },
          "404": {
            "description": "Deployment not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "default": {
            "description": "The request failed. Inspect the HTTP status and Problem Details response for recovery guidance.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Deployments"],
        "summary": "Deploy a guild",
        "description": "Creates or updates the active deployment for a guild, then records a revision.",
        "operationId": "upsert_deployment_handler",
        "parameters": [
          {
            "name": "guild_id",
            "in": "path",
            "description": "Guild ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeploymentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Deployment stored",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "default": {
            "description": "The request failed. Inspect the HTTP status and Problem Details response for recovery guidance.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/orchestrator/deployment": {
      "get": {
        "tags": ["Orchestrator"],
        "summary": "Get the orchestrator deployment",
        "description": "Returns the active trusted orchestrator deployment. Only the configured Discord operator may access this endpoint; the deployment may contain sensitive feature-policy source.",
        "operationId": "get_orchestrator_deployment",
        "responses": {
          "200": {
            "description": "Active orchestrator deployment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrchestratorDeploymentResponse"
                }
              }
            }
          },
          "404": {
            "description": "No orchestrator deployment",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "default": {
            "description": "The request failed. Inspect the HTTP status and Problem Details response for recovery guidance.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Orchestrator"],
        "summary": "Deploy the orchestrator",
        "description": "Validates and atomically replaces the trusted orchestrator isolate, persists a deployment revision, and immediately reconciles User Bot and Server Custom Bot gateways against the new feature policy. A failed deployment leaves the previous orchestrator active.",
        "operationId": "deploy_orchestrator",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeploymentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Orchestrator deployed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrchestratorDeploymentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid deployment",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "default": {
            "description": "The request failed. Inspect the HTTP status and Problem Details response for recovery guidance.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Orchestrator"],
        "summary": "Delete the orchestrator deployment",
        "description": "Removes the trusted orchestrator deployment and restores deny-by-default feature authorization. Existing User Bot and Server Custom Bot gateways are reconciled immediately without deleting their encrypted configuration.",
        "operationId": "delete_orchestrator",
        "responses": {
          "200": {
            "description": "Orchestrator deleted"
          },
          "default": {
            "description": "The request failed. Inspect the HTTP status and Problem Details response for recovery guidance.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CustomBotDeploymentResponse": {
        "type": "object",
        "required": ["entry", "created_at", "updated_at"],
        "properties": {
          "bundle": {
            "type": ["string", "null"]
          },
          "created_at": {
            "type": "string"
          },
          "entry": {
            "type": "string"
          },
          "files": {
            "type": ["array", "null"],
            "items": {
              "$ref": "#/components/schemas/DeploymentFile"
            }
          },
          "source_map": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DeploymentSourceMapFile"
              }
            ]
          },
          "updated_at": {
            "type": "string"
          }
        }
      },
...
      "DeploymentFile": {
        "type": "object",
        "required": ["path", "contents"],
        "properties": {
          "contents": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        }
      },
      "DeploymentListItem": {
        "type": "object",
        "required": ["guild_id", "entry", "created_at", "updated_at"],
        "properties": {
          "created_at": {
            "type": "string"
          },
          "entry": {
            "type": "string"
          },
          "guild_id": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        }
      },
      "DeploymentRequest": {
        "type": "object",
        "required": ["entry"],
        "properties": {
          "build_id": {
            "type": ["string", "null"]
          },
          "bundle": {
            "type": ["string", "null"]
          },
          "entry": {
            "type": "string"
          },
          "files": {
            "type": ["array", "null"],
            "items": {
              "$ref": "#/components/schemas/DeploymentFile"
            }
          },
          "source_map": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DeploymentSourceMapFile"
              }
            ]
          }
        }
      },
      "DeploymentResponse": {
        "type": "object",
        "description": "Deployment snapshot stored for a guild.",
        "required": ["guild_id", "created_at", "updated_at", "entry"],
        "properties": {
          "bundle": {
            "type": ["string", "null"],
            "description": "Bundled output when stored, if included."
          },
          "created_at": {
            "type": "string",
            "description": "Snapshot creation time in RFC3339 (UTC)."
          },
          "entry": {
            "type": "string",
            "description": "Entry file path used for the deployment."
          },
          "files": {
            "type": ["array", "null"],
            "items": {
              "$ref": "#/components/schemas/DeploymentFile"
            },
            "description": "Raw files when the deployment was uploaded as sources."
          },
          "guild_id": {
            "type": "string",
            "description": "Guild ID for the deployment."
          },
          "source_map": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DeploymentSourceMapFile",
                "description": "Source map included with the bundle, if provided."
              }
            ]
          },
          "updated_at": {
            "type": "string",
            "description": "Snapshot update time in RFC3339 (UTC)."
          }
        }
      },
...
      "DeploymentSourceMapFile": {
        "type": "object",
        "required": ["path", "contents"],
        "properties": {
          "contents": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        }
      },
...
      "OrchestratorDeploymentResponse": {
        "type": "object",
        "required": ["entry", "created_at", "updated_at"],
        "properties": {
          "bundle": {
            "type": ["string", "null"]
          },
          "created_at": {
            "type": "string"
          },
          "entry": {
            "type": "string"
          },
          "files": {
            "type": ["array", "null"],
            "items": {
              "$ref": "#/components/schemas/DeploymentFile"
            }
          },
          "source_map": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DeploymentSourceMapFile"
              }
            ]
          },
          "updated_at": {
            "type": "string"
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "required": ["type", "title", "status", "detail", "instance", "request_id"],
        "properties": {
          "detail": {
            "type": "string",
            "description": "Actionable explanation of this specific failure occurrence."
          },
          "instance": {
            "type": "string",
            "description": "Request path identifying where this failure occurred."
          },
          "request_id": {
            "type": "string",
            "description": "Correlation identifier also returned in the `X-Request-ID` header."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status code returned for this failure occurrence.",
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "Stable, human-readable summary of the failure category."
          },
          "type": {
            "type": "string",
            "description": "Stable relative URI identifying the category of failure."
          },
          "violations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProblemViolation"
            },
            "description": "Field-level validation failures, when applicable."
          }
        }
      },
      "ProblemViolation": {
        "type": "object",
        "required": ["field", "message"],
        "properties": {
          "field": {
            "type": "string",
            "description": "Identifies the invalid request field using its JSON or parameter path."
          },
          "message": {
            "type": "string",
            "description": "Explains how to correct the invalid field."
          }
        }
      },
...
    }
  },
...
}